requirements.txt
Django==3.0.7
coverage
pyaml
django-recurrence
python-decouple==3.3
pandas
djangorestframework~=3.12.0
django-cors-headers~=3.10.0
